+++ CHANGELOG.md
@@ -1,4 +1,18 @@
-## 2.0
+## 2.0.1
+
+### Changes
+
+- Changed ouput filename to `javascript-text-truncate.min.js`
+- Switched from `babel-cli` to `babel-core`
+- Updated webpack to `2.1.0-beta.25`
+- Updated development dependencies
+- Updated example
+- Fixed `.editorconfig`
+- Updated README
+
+## 2.0.0
+
+> Updated API. Read about breaking changes below.
 
 ### Bug Fixes
 
@@ -28,6 +42,6 @@ new TruncatedText({
 })
 ```
 
-## 1.0
+## 1.0.0
 
 > Initial Release